moved core of gen schedule and calc costs to spice_ev module
